feat: add product CRUD API for seller and adjust price data type to Long

+++ backend/src/main/java/com/seapedia/backend/models/Product.java
@@ -18,7 +18,7 @@ public class Product {
     private String description;
 
     @Column(nullable = false)
-    private Double price;
+    private Long price;
 
     @Column(nullable = false)
     private Integer stock;
